refactor: Move McStopArrival and related classes to 'stop' package
